feat: support validation of container images (#24)
When the flag `-run-in-container` (false by default) is set to true, we run the functions framework in a container. The function output file will be copied to the local file system for the validation. This can be used to validate container images built by our buildpacks. For example, to validate a function image called `fn-image`, run

```
./client -cmd="fn-image" -type="http" -run-in-container="true"
```