Loosen requirements on the data payload of CloudEvents.
When a legacy event is converted to a CloudEvent, we know that we want a certain
JSON payload, but we shouldn't require that payload to be formatted exactly the
same as the JSON payload in our golden files. Instead, we unmarshal both the
actual and expected values into maps, and we check that those maps are equal.