Override the expected converted Pub/Sub payload
There are multiple legacy Pub/Sub payload schemas. For example, see:

- legacy_pubsub-legacy-output.json
- pubsub_text-legacy-output.json

When converting Cloud Events to legacy events there is no information
in the CE that enables us to infer which schema to convert to.
Unfortunately, this tool assumes the converted output will always match
a corresponding unconverted payload, which isn't necessarily possible.

This commit adds a mechanism to override the exepcted output when
converting between event types and creates an override for converting
legacy_pubsub test case