feat: use temp directory for client server logs
Testing the client locally typically means testing from the root of an
FF repo, so this prevents the FF repo from needing to .gitignore the
server logs files that are produced by the client.

Also, eliminate the cross-binary dependency of the GitHub Action and the
conformance test client on knowing the logs file name by having the
client itself print the server logs on error.