fix: user non-zero ending timestamps for event data
"2006-01-02T15:04:05.000Z" is not the correct format when using the
standard RFC3339 format, trailing zeros after the decimal are normally omitted. Only
when the decimal is non-zero is the data included in the string format.

This has caused problems with using the standard time libraries in
various languages to parse the test data because standard libraries
truncate the .000, but our tests data output expects it to still be
there in the output JSON.