fix: Update buildpack-integration-test.yml (#165)
* Update buildpack-integration-test.yml

Add ${{ inputs.builder-runtime }} to the `name` for artifact upload. This will avoid the conflict in names, where e.g. php81, php82 and php83 will try to upload the artifacts with the same name and the second and later attempts will fail:

```
Run actions/upload-artifact@v4
  with:
    name: cloudevent_buildpack_integ_logs
    path: /tmp/ff_*
    retention-days: 5
    if-no-files-found: warn
    compression-level: 6
    overwrite: false
    include-hidden-files: false
With the provided path, there will be 2 files uploaded
Artifact name is valid!
Root directory input is valid!
Error: Failed to CreateArtifact: Received non-retryable error: Failed request: (409) Conflict: an artifact with this name already exists on the workflow run
```

https://github.com/GoogleCloudPlatform/functions-framework-php/actions/runs/13191006742/job/36867920300

* fix: Update buildpack-integration-test.yml

change uploaded artifact name based on whether a builder-url is specified or not